Made the RFTrigger GUI look a bit better.
Added ability to update RFTriggers

Fixed audio in concatenated shows by taking into account the length of the last movement when determining the relative timestamp of the audio.

Effect view panel no longer will update while a show is playing. The effect information is redisplayed when the show is paused. It will update as usual if the slider is used while the show is paused.

Removed rftriggerGUI while show is playing to avoid modificaitons to rftriggers that are not at the current count. (since the panel no longer updates while the show is playing.)

Did a little bit of code cleaning. Mostly making things more organized and taking out print statements. No major changes though.